[FRONT] refactor refresh session
