+++ client/src/App.js
@@ -2,15 +2,15 @@ import React from 'react';
 import {Route, Switch} from "react-router-dom";
 import './App.css';
 import Navbar from "./component/navbar/Navbar";
-import SignIn from "./component/sign-in/SignIn";
-import SignUp from "./component/sign-up/SignUp";
-import Home from "./home/Home";
+import SignIn from "./component/protected/sign-in/SignIn";
+import SignUp from "./component/protected/sign-up/SignUp";
+import Home from "./component/public/home/Home";
 import AuthContextProvider from "./context/AuthContext";
-import PrivateRoute from "./PrivateRoute";
-import UserProfile from "./component/protected/user-profile/UserProfile";
+import PrivateRoute from "./routes/PrivateRoute";
+import UserProfile from "./component/private/user-profile/UserProfile";
 import { config } from '@fortawesome/fontawesome-svg-core';
 import '@fortawesome/fontawesome-svg-core/styles.css';
-import ProtectedRoute from "./ProtectedRoute";
+import ProtectedRoute from "./routes/ProtectedRoute";
 
 const App = () => {
 
@@ -21,7 +21,6 @@ const App = () => {
       <AuthContextProvider>
         <Navbar />
         <Switch>
-
           <Route exact path='/' component={Home} />
 
           <ProtectedRoute exact path='/sign-in' component={SignIn} />
@@ -30,7 +29,6 @@ const App = () => {
           <PrivateRoute exact path="/user-profile" component={UserProfile} />
 
           <Route path="*" component={() => "404 NOT FOUND"} />
-
         </Switch>
       </AuthContextProvider>
     </div>
