+++ client/src/index.js
@@ -4,7 +4,7 @@ import './index.css';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
 import {BrowserRouter} from "react-router-dom";
-import axios from 'axios';
+import axios from "axios";
 
 axios.defaults.xsrfCookieName = 'csrf_token';
 axios.defaults.xsrfHeaderName = "X-CSRFTOKEN";
