+++ client/src/routes/PrivateRoute.js
@@ -1,6 +1,6 @@
 import React, {useContext} from 'react';
 import { Route, Redirect } from "react-router-dom";
-import {AuthContext} from "./context/AuthContext";
+import {AuthContext} from "../context/AuthContext";
 
 const PrivateRoute = ({component: Component, ...rest}) => {
